Enhance player state management: add maxHealth attribute and update related logic + adjust font format of damage

+++ src/main/java/com/demo/mota/engine/Item/GenericItem/FloorJumper.java
@@ -3,12 +3,12 @@ package com.demo.mota.engine.Item.GenericItem;
 import com.demo.mota.engine.GameEngine;
 
 public class FloorJumper extends GenericItem{
-    private short floorNumberSelected;
+    private int floorNumberSelected;
 
     public FloorJumper(String itemId, String itemName, String itemDescription,
                        long itemPrice, int itemCount,
                        boolean isStorable, boolean isConsumable,
-                       short floorNumberSelected) {
+                       int floorNumberSelected) {
         super(itemId, itemName, itemDescription, itemPrice, itemCount, isStorable, isConsumable);
         this.floorNumberSelected = floorNumberSelected;
     }
